+++ soa_loadTauSamples.m
@@ -7,11 +7,22 @@
 % License: CC-BY-4.0
 % Changes by: Johannes Keyser, please refer to changelog.md
 
-function [samples_tauA, samples_tauO] ...
+function [samples_tauA, samples_tauO, sample_size] ...
     = soa_loadTauSamples(experiment, condition)
 
 [fnameTauA, fnameTauO] = soa_getFileNames(experiment, condition);
 samples_tauA = dlmread(fnameTauA);
 samples_tauO = dlmread(fnameTauO);
+
+assert(isvector(samples_tauA) && isvector(samples_tauO), ...
+    'Saved samples are expected to be vectors.')
+
+sample_sizeA = numel(samples_tauA);
+sample_sizeO = numel(samples_tauO);
+
+assert(sample_sizeA == sample_sizeO, ...
+    'Sample sizes of tauA and tauO must be the same.')
+
+sample_size = sample_sizeA;
 
 end