+++ soa_sortMatrices.m
@@ -13,7 +13,7 @@
 function sortedMatrix = ...
     soa_sortMatrices(vectMatrix, sortIndx1, sortIndx2, sortIndx3)
 
-sortedMatrix = vectMatrix;
+sortedMatrix = nan(size(vectMatrix));
 sortedMatrix(1, :) = vectMatrix(1, sortIndx1);
 sortedMatrix(2, :) = vectMatrix(2, sortIndx2);
 sortedMatrix(3, :) = vectMatrix(3, sortIndx3);